requirements.txt
alabaster>=0.7.16
amqp>=5.2.0
asgiref>=3.7.2
async-timeout>=4.0.3
attrs>=23.2.0
autobahn>=23.6.2
Automat>=22.10.0
Babel>=2.14.0
beautifulsoup4>=4.12.3
billiard>=4.2.0
celery>=5.3.6
certifi>=2024.2.2
cffi>=1.16.0
channels>=4.0.0
channels-redis>=4.2.0
charset-normalizer>=3.3.2
click>=8.1.7
click-didyoumean>=0.3.0
click-plugins>=1.1.1
click-repl>=0.3.0
constantly>=23.10.4
coverage>=7.4.3
cryptography>=42.0.3
daphne>=4.1.0
Django>=5.0
django-channels>=0.7.0
django-ckeditor>=6.7.0
django-htmx>=1.17.2
django-js-asset>=2.2.0
django-model-utils>=4.4.0
django-notifications-hq>=1.8.3
djangorestframework>=3.14.0
djhtml>=3.0.6
docutils>=0.20.1
drf-yasg>=1.21.7
exceptiongroup>=1.2.0
furo>=2024.1.29
gunicorn>=21.2.0
hyperlink>=21.0.0
idna>=3.6
imagesize>=1.4.1
incremental>=22.10.0
inflection>=0.5.1
iniconfig>=2.0.0
Jinja2>=3.1.3
jsonfield>=3.1.0
kombu>=5.3.5
MarkupSafe>=2.1.5
msgpack>=1.0.7
oauthlib>=3.2.2
packaging>=23.2
pillow>=10.2.0
pluggy>=1.4.0
prompt-toolkit>=3.0.43
pyasn1>=0.5.1
pyasn1-modules>=0.3.0
pycparser>=2.21
Pygments>=2.17.2
pyOpenSSL>=24.0.0
pytest>=8.0.2
pytest-django>=4.8.0
python-dateutil>=2.8.2
pytz>=2023.4
PyYAML>=6.0.1
redis>=5.0.1
requests>=2.31.0
requests-oauthlib>=1.3.1
service-identity>=24.1.0
six>=1.16.0
snowballstemmer>=2.2.0
soupsieve>=2.5
Sphinx>=7.2.6
sphinx-basic-ng>=1.0.0b2
sphinx-copybutton>=0.5.2
sphinxcontrib-applehelp>=1.0.8
sphinxcontrib-devhelp>=1.0.6
sphinxcontrib-htmlhelp>=2.0.5
sphinxcontrib-jsmath>=1.0.1
sphinxcontrib-qthelp>=1.0.7
sphinxcontrib-serializinghtml>=1.1.10
sqlparse>=0.4.4
swapper>=1.3.0
tomli>=2.0.1
Twisted>=23.10.0
txaio>=23.1.1
typing_extensions>=4.9.0
tzdata>=2023.4
uritemplate>=4.1.1
urllib3>=2.2.0
vine>=5.1.0
wcwidth>=0.2.13
websockets>=12.0
whitenoise>=6.6.0
zope.interface>=6.2
